feat: play FLAC and SACD disc images, locking the DAC to each rate
FLAC is decoded to interleaved PCM and sent at the file's own sample rate
and width: the device is claimed and its stream format set to that rate, so
nothing resamples or mixes. Codes are left-justified into the same 24-bit
word DoP frames use, which is a shift by a power of two and so exact. Seeks
use the file's seek table where it has one.

A SACD image opens as the list of tracks the disc holds, and the DST the
audio is compressed with is decoded here: the arithmetic coder, prediction
filters and probability tables are read from each frame and reset at its
start. Decoding runs at about twenty times real time per core for stereo
DSD64. Seeks halve the track's sectors by the timecode each audio frame
carries rather than decoding the distance.

The queue between the reader and the render callback now carries carrier
words for either kind: DoP payloads, which the callback still wraps in
alternating markers, or PCM codes written out as they stand. The native USB
DSD path is unchanged.

+++ src/dop.rs
@@ -43,7 +43,10 @@ pub const fn split_word(word: i32) -> (u8, u16) {
 
 /// Interleave planar MSB-first DSD into DoP payloads: two DSD bytes per frame per channel.
 /// A plane of odd length pairs its final byte with DSD silence.
-pub fn pack_planes(planes: &[&[u8]], out: &mut Vec<u16>) -> usize {
+///
+/// Payloads ride in the low sixteen bits of the carrier word the queue carries; the marker
+/// that completes the DoP frame is added by the render callback, which owns the alternation.
+pub fn pack_planes(planes: &[&[u8]], out: &mut Vec<i32>) -> usize {
     let Some(first) = planes.first() else {
         return 0;
     };
@@ -56,7 +59,7 @@ pub fn pack_planes(planes: &[&[u8]], out: &mut Vec<u16>) -> usize {
                 .get(frame * 2 + 1)
                 .copied()
                 .unwrap_or(DSD_SILENCE_BYTE);
-            out.push(u16::from_be_bytes([high, low]));
+            out.push(i32::from(u16::from_be_bytes([high, low])));
         }
     }
     frames
@@ -74,7 +77,7 @@ mod tests {
         for frame in payloads.chunks(planes.len()) {
             let byte = marker.next();
             for payload in frame {
-                words.push(word(byte, *payload));
+                words.push(word(byte, *payload as u16));
             }
         }
         words
